medvbotek/requirements.txt
python-telegram-bot==20.7
requests
pymongo[srv]
python-dotenv
aiohttp